rename framework: msx2 -> MSX2
