+++ msx2-ios/MSX2.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "self:">
+      location = "self:/Users/suzukiplan/dev/picopico/msx2-osx/msx2-ios/MSX2.xcodeproj">
    </FileRef>
 </Workspace>
