+++ msx2-ios/MSX2/MSX2Core.h
@@ -25,7 +25,7 @@
  * -----------------------------------------------------------------------------
  */
 #import <Foundation/Foundation.h>
-#import <msx2/msx2def.h>
+#import <MSX2/msx2def.h>
 
 NS_ASSUME_NONNULL_BEGIN
