+++ msx2-ios/MSX2/sha1.hpp
